directory commit

+++ tools.descartes.dlim.generator.tests/src/tools/descartes/dlim/tests/DlimTestUtils.java
@@ -53,7 +53,9 @@ public final class DlimTestUtils {
 				if (f.isDirectory()) {
 					cleanDirectory(f);
 				}
-				f.delete();
+				if (!f.getName().equals(".gitignore")) {
+					f.delete();
+				}
 			}
 		}
 	}
